fix: 프로덕션 환경에서만 KisToken 삭제 및 조회 수행하도록 수정

+++ apps/peek/src/module/schedule/kis-token/kis-token-schedule.service.ts
@@ -19,8 +19,10 @@ export class KisTokenScheduleService implements OnModuleInit {
 
   async onModuleInit() {
     try {
-      await this._deleteKisToken();
-      await this._getKisTokenSchedule();
+      if (this.configService.get('NODE_ENV') === 'production') {
+        await this._deleteKisToken();
+        await this._getKisTokenSchedule();
+      }
     } catch (error) {
       console.error('KisTokenScheduleService onModuleInit 에러:', error);
     }
